refactor: TSify 1470-Shuffle The Array
