--- Algorithm/1470-ShuffleTheArray/shuffle.js
@@ -1,17 +1,0 @@
-/**
- * @param {number[]} nums
- * @param {number} n
- * @return {number[]}
- */
-const shuffle = function (nums, n) {
-  const result = [];
-
-  for (let i = 0; i < n; i++) {
-    result.push(nums[i]);
-    result.push(nums[n + i]);
-  }
-
-  return result;
-};
-
-module.exports = { shuffle };
